requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
blinker==1.9.0
Brotli==1.1.0
celery==5.4.0
certifi==2025.1.31
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
ConfigArgParse==1.7
Django==5.1.6
django-environ==0.12.0
djangorestframework==3.15.2
Flask==3.1.0
flask-cors==5.0.1
Flask-Login==0.6.3
gevent==24.11.1
geventhttpclient==2.3.3
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
kombu==5.4.2
locust==2.33.0
MarkupSafe==3.0.2
msgpack==1.1.0
packaging==24.2
prompt_toolkit==3.0.50
psutil==7.0.0
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
pyzmq==26.2.1
redis==5.2.1
requests==2.32.3
setuptools==75.8.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
zope.event==5.0
zope.interface==7.2
